Finalize dev release tarballs used during deploy and cats
* This ensures that releases are published using the same artifact that
made it through the pipeline. Unfortunately this means that those
artifacts must exist as resources, so this will cost us more in s3
storage.
* This commit also splits out publishing releases into two jobs per
buildpack. The first job determines if the version has already been
published. If it hasn't, it updates a file in S3 which triggers the
actual publish job. This avoids the problem of having red jobs when
ship-it is triggered but some buildpacks don't have new versions.

[#170983713]